Produce 2-player round robin matches correctly
The old code for the round robin would only work if there were only
two players in the tournament because it just randomized them and
assumed that create_match would grab the first two players.  This code
explicitly goes through a double-nested loop to produce all possible
pairs of players and does the right thing.