Incorporate networked communication for players
Modify the AIContest so that it will work better with the game contest
server.  This was a major refactor to change from cout/cerr output and
direct calls to the player to become output sent to the manager via
the network and networked communication between the referee and
players via proxies.

Almost everything is complete except figuring out how to include the
proper player code in the proxy.