Separate player define from makefile
In order to let the student define their player name and classes, we
will include a file named cpp-player.mk in their submission.  That
file can define the player and any other compilation instructions
necessary to complete their player.  We will include the makefile as
the instructor, which includes the cpp-player.mk file as part of the
build process so that we can incorporate all the student's needs while
still making the proxy based on our code--not theirs.