Handle broken sockets & code cleanup
In an effort to chase down a bug that was exhibited when a player
socket was closed too early, I did some code cleanup mostly with
variables that were not being used any longer by the gcs referee.
Additionally, extra cerr output was removed since it will not be
recorded by the server.

Once the bug was narrowed down, it was easy to see that the problem
was that communication with a player that was finished was causing a
SIGPIPE instead of just returning an error.  Installing a signal
handler that ignored the error returned the program to expected state
and things worked correctly.

This commit includes both the fix and code cleanup because it is
slightly intertwined.