Log ship placements
In order for the replay functionality to be able to handle a kill, we
need to pass additional information about the kill.  Using the
existing architecture, either

    1. we need to have multiple moves which indicate all the locations
       for the ship that was killed, or
    2. we need to add moves that show the placement of ships.

Option #1 is less preferred because the interface assumes each move is
meaningful and predictable.  We don't know how many ship elements will
be generated by a single kill shot.

Option #2 is more preferred because it allows us to make the UI better
because we can do cool things like turn on and off a display of where
the ships are while they are being shot at.  Without option #2, we
only know where ships are when they've been hit.  Since the players
don't have access to the logs during run time, this is a better
choice.