Commit initial replay prototype
This is the first cut at the prototype code.  It needs to be tested
and run on the browser, but that requires using the server, etc.  This
is a commit to get it in git so it can be pulled and worked on from
the server.  It is very possible that there are syntax errors in
addition to any logic errors that inevitably exist throughout the
code.