Handle timeout gracefully
This code is supposed to keep track of the timeout given to the
referee by the manager so that it can terminate the game based on who
we were last waiting for a message.  This is not perfect if both
players are making progress, but works correctly if one has an
infinite loop.  If we want to put tighter constraints on each player,
we will probably have to insert multiple alarm() calls (or something
even more fine grained) so that we can attribute the timeout to the
slow player and not to the last player that got caught when the alarm
is triggered.