Make new directory for Guess W example
The examples directory was getting big enough and has enough examples
that the simplistic guess W example was confusing with the other
directories with specific examples.  This commit gives guess W its own
directory to clearly delineate what belongs to that example and how it
is/isn't related to any other examples.