Move Timecop.freeze out of test configuration
Libby's work to add Javascript to some tests conflicted with Timecop
explicitly.  She just commented it out to let her tests run, but this
is the submission that makes sure that only the tests that need
Timecop actually use it.  There are now only two instances of Timecop:
one each in spec/modes/contest_spec.rb and spec/models/match_spec.rb.

The other big set of changes was to be able to handle dynamic dates and
times without the help of Timecop anymore.  By far, the biggest set of
changes were handling minutes in times since we only display select boxes
with multiples of 5 for the minutes.  The tests now embed that assumption
deeply (see spec/support/dates.rb and spec/support.forms.rb).

All tests properly pass with the exception of the notorious 4 that have been
failing all j-term long.