Update to Rails 5.1
This mostly works *EXCEPT* for the validations that use the
validates_timeliness gem.  That gem has not been updated for Rails 5.0 and so
it breaks hard.  I'll have to fix that going forward, but all the other tests
pass and so they should be able to be used in development while this last
fix gets worked on.