Compile correctly with PlayerConnection
The player.cpp was missing the last_wait global variable being defined
somewhere.  This makes it defined in the proper location.  It isn't used
in the player, but is in the referee and is a cheap way to keep track of
which player "caused" the timeout.  We don't do anything with this
information, so far, so we just need to get the player to properly
compile.

Also, slightly fix up the Makefile to get the rest of things compiling
correctly.