Run the specified number of rounds
For round robin tournaments, we were running double the number of
specified rounds because we would do both match-ups for each player.
This makes sure that we play them the correct number of times and
roughly half with each player as the first player and the others as the
second player.